Tableau Dashboards of EDA process
During the Exploratory Data Analysis, I was able to create Tableau dashboards that were effective in giving the clinicians at Bruner an opportunity to look at their patient population with a more hands-on approach. The goal of these was to answer some initial questions that we had, but also possibly uncover other questions which myself and the group of clinicians I worked with the opportunity to ask more in-depth questions about the patients they see on a regular basis.