Alteryx Workflows
I had two different workflows that I created for this patient data. Bruner_Clinic_Workflow extracts cleans and manipulates patient data that only deals with patients that have either Hypertension, Diabetes or Obesity in the last 3 years. The BC Workflows for SLM are a patient data pull over the last year that brings in patients no matter their diagnosis, but adds in 3 separate fields: one to denote if a patient has hypertension or not, one that classifies if they have diabetes, and a final one that denotes if they have obesity.